client: move UiAction out of keyboard input module
