+++ rigflow-client/src/app/mod.rs
@@ -7,3 +7,4 @@ pub mod state;
 pub mod stats;
 pub mod spectrum_utils;
 pub mod waterfall;
+pub mod ui_action;
