+++ rigflow-client/src/app/ui_action.rs
@@ -1,0 +1,16 @@
+#[derive(Debug, Clone)]
+pub enum UiAction {
+    SetTargetFrequency(f32),
+    SetCenterFrequency(f32),
+    SetDemodMode(&'static str),
+    SetSideband(&'static str),
+    SetSsbPitch(f32),
+
+    ToggleRigflowServerMenu,
+    FocusRigflowServerIpField,
+    ConnectToRigflowServer,
+    DisconnectFromRigflowServer,
+
+    CycleLicenseForward,
+    CycleLicenseBackward,
+}
